ajust orders

+++ Modules/Events/Database/Migrations/2019_12_10_212140_create_events_table.php
@@ -54,9 +54,11 @@ class CreateEventsTable extends Migration
             $table->foreign('event_id')->references('id')->on('events')->onDelete('cascade');
         });
 
-        Schema::table('transactions', function (Blueprint $table) {
-            $table->foreign('event_id')->references('id')->on('events')->onDelete('cascade');
-        });
+        if (Schema::hasTable('transactions') && !Schema::hasColumn('transactions', 'event_id')) {
+            Schema::table('transactions', function (Blueprint $table) {
+                $table->foreign('event_id')->references('id')->on('events')->onDelete('cascade');
+            });
+        }
     }
 
     /**
